# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Leoflow Public API
  description: |
    Airflow-3.2.x-compatible subset of the public API.
    This is the surface consumed by the Airflow UI when configured to talk to a Leoflow Control Plane.
  version: "1.0.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v2/users:
    post:
      summary: Create a user
      description: |
        Admin-only. Creates a control-plane account with the given email and
        password and grants at most one role. The password is write-only and is
        never returned. Requires the admin:tenant permission.
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Invalid input (missing fields or unknown role)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: A user with this email already exists in the tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: integer }
        instance: { type: string }
...
    CreateUserRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
        role:
          type: string
          description: >-
            Name of an existing role to grant (e.g. "admin"). Omit to grant no
            role — the most restrictive default; the user then holds no
            permissions until an admin grants one.
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string }
        email: { type: string }
        role: { type: string }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...